@@ -4,7 +4,11 @@ module Commentable
   included do
     attr_readonly :comments_count
 
-    has_many :comments, :class_name => 'Message', :as => :target, :order => 'created_at DESC'
+    has_many :comments, :class_name => 'Message', :as => :target
+
+    # TODO
+    # :order => 'created_at DESC'
+
   end
 
   def add_comment(user, content, options = {})
