@@ -1,19 +1,15 @@
 require File.expand_path('../boot', __FILE__)
 
 # Pick the frameworks you want:
-require "active_record/railtie"
-require "action_controller/railtie"
-require "action_mailer/railtie"
-require "active_resource/railtie"
-require "sprockets/railtie"
+# require "active_record/railtie"
+# require "action_controller/railtie"
+# require "action_mailer/railtie"
+# require "active_resource/railtie"
+# require "sprockets/railtie"
+require 'rails/all'
 # require "rails/test_unit/railtie"
 
-if defined?(Bundler)
-  # If you precompile assets before deploying to production, use this line
-  Bundler.require(*Rails.groups(:assets => %w(development test)))
-  # If you want your assets lazily compiled in production, use this line
-  # Bundler.require(:default, :assets, Rails.env)
-end
+Bundler.require(*Rails.groups)
 
 module AngelNest
   class Application < Rails::Application
@@ -58,7 +54,7 @@ module AngelNest
     # This will create an empty whitelist of attributes available for mass-assignment for all models
     # in your app. As such, your models will need to explicitly whitelist or blacklist accessible
     # parameters by using an attr_accessible or attr_protected declaration.
-    config.active_record.whitelist_attributes = true
+    # config.active_record.whitelist_attributes = true
 
     # Enable the asset pipeline
     config.assets.enabled = true
@@ -75,3 +71,7 @@ end
 
 require Rails.root + 'lib/monkey_patches/array'
 require Rails.root + 'lib/monkey_patches/warden'
+# require Rails.root + 'lib/monkey_patches/duration'
+# require Rails.root + 'lib/monkey_patches/enumerable'
+# require Rails.root + 'lib/monkey_patches/time_in_integer'
+# require Rails.root + 'lib/monkey_patches/time_in_numeric'
